Major pipeline header cleanup.
