+++ CMakeLists.txt
@@ -156,16 +156,18 @@ set(RENDERER_SOURCES
         src/renderer/environment/environment.h
         src/renderer/environment/environment_constants.h
         src/renderer/environment/environment_types.h
-        src/renderer/pipelines/deferred_mrt/deferred_mrt.cpp
-        src/renderer/pipelines/deferred_mrt/deferred_mrt.h
-        src/renderer/pipelines/deferred_resolve/deferred_resolve.cpp
-        src/renderer/pipelines/deferred_resolve/deferred_resolve.h
+        src/renderer/pipelines/deferred_mrt/deferred_mrt_pipeline.cpp
+        src/renderer/pipelines/deferred_mrt/deferred_mrt_pipeline.h
+        src/renderer/pipelines/deferred_mrt/deferred_mrt_pipeline_types.h
+        src/renderer/pipelines/deferred_resolve/deferred_resolve_pipeline.cpp
+        src/renderer/pipelines/deferred_resolve/deferred_resolve_pipeline.h
+        src/renderer/pipelines/deferred_resolve/deferred_resolve_pipeline_types.h
         src/renderer/pipelines/environment/environment_pipeline.cpp
         src/renderer/pipelines/environment/environment_pipeline.h
         src/renderer/pipelines/post_process/post_process_pipeline.cpp
         src/renderer/pipelines/post_process/post_process_pipeline.h
-        src/renderer/pipelines/visibility_pass/visibility_pass.cpp
-        src/renderer/pipelines/visibility_pass/visibility_pass.h
+        src/renderer/pipelines/visibility_pass/visibility_pass_pipeline.cpp
+        src/renderer/pipelines/visibility_pass/visibility_pass_pipeline.h
         src/renderer/pipelines/temporal_antialiasing_pipeline/temporal_antialiasing_pipeline.cpp
         src/renderer/pipelines/temporal_antialiasing_pipeline/temporal_antialiasing_pipeline.h
         src/renderer/pipelines/basic_compute/basic_compute_pipeline.cpp
@@ -188,7 +190,6 @@ set(RENDERER_SOURCES
         src/renderer/descriptor_buffer/descriptor_buffer_uniform.cpp
         src/renderer/descriptor_buffer/descriptor_buffer_uniform.h
         src/renderer/assets/render_object/render_object_constants.h
-        src/renderer/post_process/post_process_types.h
 )
 
 
@@ -247,6 +248,18 @@ set(TEMP_SOURCES
         src/renderer/pipelines/transparent_pipeline/transparent_pipeline.cpp
         src/renderer/pipelines/transparent_pipeline/transparent_pipeline.h
         src/engine_constants.h
+        src/renderer/pipelines/contact_shadow/contact_shadows_pipeline.cpp
+        src/renderer/pipelines/contact_shadow/contact_shadows_pipeline.h
+        src/renderer/pipelines/contact_shadow/contact_shadows_pipeline_types.h
+
+        src/renderer/pipelines/environment/environment_pipeline_types.h
+        src/renderer/pipelines/basic_render/basic_render_pipeline_types.h
+        src/renderer/pipelines/post_process/post_process_pipeline_types.h
+        src/renderer/pipelines/temporal_antialiasing_pipeline/temporal_antialiasing_pipeline_types.h
+        src/renderer/pipelines/terrain/terrain_pipeline_types.h
+        src/renderer/pipelines/transparent_pipeline/transparent_pipeline_types.h
+        src/renderer/pipelines/basic_compute/basic_compute_pipeline_types.h
+        src/renderer/pipelines/visibility_pass/visibility_pass_pipeline_types.h
 
 )
