+++ assets/settings.willengine
@@ -19,20 +19,20 @@
     },
     "cameraProperties": {
         "position": [
-            -0.008705933578312397,
-            71.13257598876953,
-            0.5683935880661011
+            62.77393341064453,
+            81.94156646728516,
+            -87.18507385253906
         ],
         "rotation": [
-            0.1442754864692688,
-            -0.015505615621805191,
-            -0.0022719609551131725,
-            -0.9894134402275085
+            -0.06119464710354805,
+            -0.3182604908943176,
+            0.020581098273396492,
+            -0.9458022713661194
         ],
         "fov": 1.3089969158172607,
         "aspectRatio": 1.7777777910232544,
         "nearPlane": 1000.0,
         "farPlane": 0.10000000149011612
     },
-    "environmentMapIndex": 3
+    "environmentMapIndex": 1
 }
