+++ src/renderer/imgui_wrapper.cpp
@@ -15,6 +15,7 @@
 #include "lighting/ambient_occlusion/ground_truth/ground_truth_ambient_occlusion.h"
 #include "lighting/shadows/cascaded_shadow_map.h"
 #include "lighting/shadows/shadow_constants.h"
+#include "pipelines/post_process/post_process_pipeline_types.h"
 #include "src/core/engine.h"
 #include "src/core/time.h"
 #include "src/core/camera/free_camera.h"
@@ -232,38 +233,38 @@ void ImguiWrapper::imguiInterface(Engine* engine)
                     ImGui::Checkbox("Disable Jitter", &engine->bDisableJitter);
                 }
 
-                static bool tonemapping = (engine->postProcessData & post_process::PostProcessType::Tonemapping) !=
-                                          post_process::PostProcessType::None;
+                static bool tonemapping = (engine->postProcessData & post_process_pipeline::PostProcessType::Tonemapping) !=
+                                          post_process_pipeline::PostProcessType::None;
                 if (ImGui::CollapsingHeader("Tonemapping", ImGuiTreeNodeFlags_DefaultOpen)) {
                     if (ImGui::Checkbox("Enable Tonemapping", &tonemapping)) {
                         if (tonemapping) {
-                            engine->postProcessData |= post_process::PostProcessType::Tonemapping;
+                            engine->postProcessData |= post_process_pipeline::PostProcessType::Tonemapping;
                         }
                         else {
-                            engine->postProcessData &= ~post_process::PostProcessType::Tonemapping;
+                            engine->postProcessData &= ~post_process_pipeline::PostProcessType::Tonemapping;
                         }
                     }
                 }
                 if (ImGui::CollapsingHeader("Sharpening", ImGuiTreeNodeFlags_DefaultOpen)) {
-                    static bool sharpening = (engine->postProcessData & post_process::PostProcessType::Sharpening) !=
-                                             post_process::PostProcessType::None;
+                    static bool sharpening = (engine->postProcessData & post_process_pipeline::PostProcessType::Sharpening) !=
+                                             post_process_pipeline::PostProcessType::None;
                     if (ImGui::Checkbox("Enable Sharpening", &sharpening)) {
                         if (sharpening) {
-                            engine->postProcessData |= post_process::PostProcessType::Sharpening;
+                            engine->postProcessData |= post_process_pipeline::PostProcessType::Sharpening;
                         }
                         else {
-                            engine->postProcessData &= ~post_process::PostProcessType::Sharpening;
+                            engine->postProcessData &= ~post_process_pipeline::PostProcessType::Sharpening;
                         }
                     }
                 }
                 if (ImGui::CollapsingHeader("FXAA", ImGuiTreeNodeFlags_DefaultOpen)) {
-                    static bool fxaa = (engine->postProcessData & post_process::PostProcessType::FXAA) != post_process::PostProcessType::None;
+                    static bool fxaa = (engine->postProcessData & post_process_pipeline::PostProcessType::FXAA) != post_process_pipeline::PostProcessType::None;
                     if (ImGui::Checkbox("Enable FXAA", &fxaa)) {
                         if (fxaa) {
-                            engine->postProcessData |= post_process::PostProcessType::FXAA;
+                            engine->postProcessData |= post_process_pipeline::PostProcessType::FXAA;
                         }
                         else {
-                            engine->postProcessData &= ~post_process::PostProcessType::FXAA;
+                            engine->postProcessData &= ~post_process_pipeline::PostProcessType::FXAA;
                         }
                     }
                 }
@@ -1077,61 +1078,6 @@ void ImguiWrapper::imguiInterface(Engine* engine)
     ImGui::End();
 
     if (ImGui::Begin("Discardable Debug")) {
-        if (transparentDebugTextureImguiId == VK_NULL_HANDLE) {
-            if (engine->transparentPipeline->debugImage.image != VK_NULL_HANDLE) {
-                transparentDebugTextureImguiId = ImGui_ImplVulkan_AddTexture(
-                    engine->resourceManager->getDefaultSamplerNearest(),
-                    engine->transparentPipeline->debugImage.imageView,
-                    VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
-            }
-        }
-
-        ImGui::BeginChild("TransparentDebugPreview", ImVec2(0, 0), false, ImGuiWindowFlags_None);
-
-        if (transparentDebugTextureImguiId == VK_NULL_HANDLE) {
-            ImGui::TextColored(ImVec4(1.0f, 0.3f, 0.3f, 1.0f), "Debug texture not available.");
-        }
-        else {
-            // Calculate best fit size
-            float maxSize = ImGui::GetContentRegionAvail().x;
-            maxSize = glm::min(maxSize, 1024.0f);
-
-            VkExtent3D imageExtent = engine->transparentPipeline->debugImage.imageExtent;
-            float width = std::min(maxSize, static_cast<float>(imageExtent.width));
-            float aspectRatio = static_cast<float>(imageExtent.width) / static_cast<float>(imageExtent.height);
-            float height = width / aspectRatio;
-
-            ImGui::Image(reinterpret_cast<ImTextureID>(transparentDebugTextureImguiId), ImVec2(width, height));
-
-            if (ImGui::Button("Save Transparency Debug Image")) {
-                if (file::getOrCreateDirectory(file::imagesSavePath)) {
-                    const std::filesystem::path path = file::imagesSavePath / "transparent_debug.png";
-
-                    vk_helpers::saveImageRGBA16SFLOAT(
-                        *engine->resourceManager,
-                        *engine->immediate,
-                        engine->transparentPipeline->debugImage,
-                        VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL,
-                        VK_IMAGE_ASPECT_COLOR_BIT,
-                        path.string().c_str(),
-                        false
-                    );
-
-                    ImGui::OpenPopup("SaveConfirmation");
-                }
-            }
-
-            if (ImGui::BeginPopupModal("SaveConfirmation", NULL, ImGuiWindowFlags_AlwaysAutoResize)) {
-                ImGui::Text("Image saved to %s/gtao_debug.png", file::imagesSavePath.string().c_str());
-                if (ImGui::Button("OK", ImVec2(120, 0))) {
-                    ImGui::CloseCurrentPopup();
-                }
-                ImGui::EndPopup();
-            }
-
-            ImGui::EndChild();
-        }
-
         if (postProcessOutputImguiId == VK_NULL_HANDLE) {
             if (engine->postProcessOutputBuffer.imageView != VK_NULL_HANDLE) {
                 postProcessOutputImguiId = ImGui_ImplVulkan_AddTexture(
