+++ src/renderer/imgui_wrapper.h
@@ -85,8 +85,6 @@ private:
     bool showGtaoDebugPreview = false;
     VkDescriptorSet aoDebugTextureImguiId{VK_NULL_HANDLE};
 
-    VkDescriptorSet transparentDebugTextureImguiId{VK_NULL_HANDLE};
-
     VkDescriptorSet postProcessOutputImguiId{VK_NULL_HANDLE};
 };
 }
