+++ src/renderer/pipelines/basic_render/basic_render_pipeline.h
@@ -4,34 +4,20 @@
 
 #ifndef BASIC_RENDER_PIPELINE_H
 #define BASIC_RENDER_PIPELINE_H
-#include "src/renderer/resource_manager.h"
-#include "src/renderer/vulkan_context.h"
+
+#include <volk/volk.h>
+
 #include "src/renderer/descriptor_buffer/descriptor_buffer_sampler.h"
 
-
-namespace will_engine::basic_render
+namespace will_engine
 {
-struct RenderDescriptorInfo
-{
-    VkSampler sampler{VK_NULL_HANDLE};
-    VkImageView texture{VK_NULL_HANDLE};
-};
+class ResourceManager;
+}
 
-struct RenderDrawInfo
+namespace will_engine::basic_render_pipeline
 {
-    VkExtent2D renderExtent{};
-    VkImageView drawImage{VK_NULL_HANDLE};
-    VkImageView depthImage{VK_NULL_HANDLE};
-    VkDescriptorBufferBindingInfoEXT sceneDataBinding{};
-    VkDeviceSize sceneDataOffset{0};
-    int32_t currentFrame{};
-};
-
-struct RenderPushConstant
-{
-    int32_t currentFrame;
-};
-
+struct RenderDrawInfo;
+struct RenderDescriptorInfo;
 
 class BasicRenderPipeline
 {
