+++ src/renderer/pipelines/contact_shadow/contact_shadows_pipeline.cpp
@@ -1,0 +1,13 @@
+//
+// Created by William on 2025-04-14.
+//
+
+#include "contact_shadows_pipeline.h"
+
+namespace will_engine::contact_shadows_pipeline
+{
+ContactShadowsPipeline::ContactShadowsPipeline(ResourceManager& resourceManager) : resourceManager(resourceManager)
+{
+
+}
+}
