+++ src/renderer/pipelines/post_process/post_process_pipeline.h
@@ -4,33 +4,20 @@
 
 #ifndef POST_PROCESS_PIPELINE_H
 #define POST_PROCESS_PIPELINE_H
-#include "src/renderer/resource_manager.h"
-#include "src/renderer/post_process/post_process_types.h"
 
+#include <volk/volk.h>
+
+#include "src/renderer/descriptor_buffer/descriptor_buffer_sampler.h"
+
+namespace will_engine
+{
+class ResourceManager;
+}
 
 namespace will_engine::post_process_pipeline
 {
-struct PostProcessDescriptor
-{
-    VkImageView inputImage;
-    VkImageView outputImage;
-    VkSampler sampler;
-};
-
-struct PostProcessPushConstants
-{
-    int32_t width;
-    int32_t height;
-    uint32_t postProcessFlags;
-    int32_t padding;
-};
-
-struct PostProcessDrawInfo
-{
-    post_process::PostProcessType postProcessFlags{post_process::PostProcessType::ALL};
-    VkDescriptorBufferBindingInfoEXT sceneDataBinding{};
-    VkDeviceSize sceneDataOffset{0};
-};
+struct PostProcessDrawInfo;
+struct PostProcessDescriptor;
 
 class PostProcessPipeline
 {
