+++ src/renderer/pipelines/terrain/terrain_pipeline.h
@@ -4,35 +4,20 @@
 
 #ifndef TERRAIN_PIPELINE_H
 #define TERRAIN_PIPELINE_H
-#include "src/core/game_object/terrain.h"
-#include "src/renderer/renderer_constants.h"
-#include "src/renderer/resource_manager.h"
 
+#include <volk/volk.h>
+
+#include "src/renderer/descriptor_buffer/descriptor_buffer_sampler.h"
+
+namespace will_engine
+{
+class ResourceManager;
+}
 
 namespace will_engine::terrain
 {
+struct TerrainDrawInfo;
 class TerrainChunk;
-
-struct TerrainPushConstants
-{
-    float tessLevel;
-};
-
-struct TerrainDrawInfo
-{
-    bool bClearColor{true};
-    bool bDrawLinesOnly{false};
-    int32_t currentFrameOverlap{0};
-    glm::vec2 viewportExtents{RENDER_EXTENT_WIDTH, RENDER_EXTENT_HEIGHT};
-    const std::unordered_set<ITerrain*>& terrains;
-    VkImageView normalTarget{VK_NULL_HANDLE};
-    VkImageView albedoTarget{VK_NULL_HANDLE};
-    VkImageView pbrTarget{VK_NULL_HANDLE};
-    VkImageView velocityTarget{VK_NULL_HANDLE};
-    VkImageView depthTarget{VK_NULL_HANDLE};
-    VkDescriptorBufferBindingInfoEXT sceneDataBinding{};
-    VkDeviceSize sceneDataOffset{0};
-};
 
 class TerrainPipeline
 {
@@ -40,8 +25,6 @@ public:
     explicit TerrainPipeline(ResourceManager& resourceManager);
 
     ~TerrainPipeline();
-
-    //void setupDescriptorBuffer(const TerrainDescriptor& descriptor);
 
     void draw(VkCommandBuffer cmd, const TerrainDrawInfo& drawInfo) const;
